Refactor cli command to be its own command
