+++ tools/cli/src/cli/lib.rs
@@ -1,3 +1,4 @@
+#![feature(future_join)]
 #![deny(warnings)]
 #![warn(unused_extern_crates)]
 #![deny(clippy::todo)]
@@ -33,6 +34,7 @@ mod raw;
 mod recycle;
 mod serviceaccount;
 mod session;
+mod graph;
 mod synch;
 mod system_config;
 mod webauthn;
@@ -168,6 +170,7 @@ impl KanidmClientOpt {
             KanidmClientOpt::Group { commands } => commands.debug(),
             KanidmClientOpt::Person { commands } => commands.debug(),
             KanidmClientOpt::ServiceAccount { commands } => commands.debug(),
+            KanidmClientOpt::Graph(gopt) => gopt.debug(),
             KanidmClientOpt::System { commands } => commands.debug(),
             KanidmClientOpt::Recycle { commands } => commands.debug(),
             KanidmClientOpt::Version {} => {
@@ -188,6 +191,7 @@ impl KanidmClientOpt {
             KanidmClientOpt::Person { commands } => commands.exec().await,
             KanidmClientOpt::ServiceAccount { commands } => commands.exec().await,
             KanidmClientOpt::Group { commands } => commands.exec().await,
+            KanidmClientOpt::Graph(gops) => gops.exec().await,
             KanidmClientOpt::System { commands } => commands.exec().await,
             KanidmClientOpt::Recycle { commands } => commands.exec().await,
             KanidmClientOpt::Version {} => (),
