+++ tools/cli/src/opt/kanidm.rs
@@ -236,9 +236,6 @@ pub enum GroupOpt {
         #[clap(flatten)]
         copt: CommonOpt,
     },
-    /// Prints graphviz dot file of all groups
-    #[clap(name = "graph")]
-    Graph(GraphCommonOpt),
     /// Delete a group
     #[clap(name = "delete")]
     Delete(Named),
@@ -285,7 +282,8 @@ pub enum GroupOpt {
 #[derive(Clone, Debug, ValueEnum)]
 pub enum GraphType {
     Graphviz,
-    Mermaid
+    Mermaid,
+    MermaidElk
 }
 #[derive(Clone, Copy, PartialEq, Eq, Hash, Debug, ValueEnum)]
 pub enum ObjectType {
@@ -1352,6 +1350,9 @@ pub enum KanidmClientOpt {
         #[clap(subcommand)]
         commands: ServiceAccountOpt,
     },
+    /// Prints graphviz dot file of all groups
+    #[clap(name = "graph")]
+    Graph(GraphCommonOpt),
     /// System configuration operations
     System {
         #[clap(subcommand)]
